判断更新

+++ privilege-service/src/main/java/cn/com/open/opensass/privilege/api/UserRoleGetPrivilegeController.java
@@ -163,7 +163,7 @@ public class UserRoleGetPrivilegeController extends BaseControllerUtil {
                     //组织机构版本相同，则走缓存
                     JSONObject userData = JSONObject.fromObject(res);
                     Integer version = (Integer) userData.get("groupVersion");
-                    if (version != null && version == groupVersion) {
+                    if (version != null && version.intValue() == groupVersion.intValue()) {
                         writeJsonString(response, res);
                         return;
                     }
